Measure async speed, instead of synchronous
The thunks measurement was based on synchronous (recursive) callbacks into co, but node I/O callbacks are never invoked until *after* the API call returns.  This change gives a (somewhat) better idea of actual callback performance.  process.nextTick is used in order to minimize the impact of delaying the callback (so we're not measuring node's I/O loop), and only the thunk benchmark is changed, because including context switching overhead in the other benchmarks would effectively measure it *twice*.